Fixed error saving the OCR result as a txt file to history. Doing so I encountered the following errors which I fixed. In doing so I modified the classes FireBaseHistoryRepository, Config, MainAppUI,
First error was that we were using the wrong database we needed a realtime database instead. However for the rules I made read and write both True which means anyone can view and modify/delete it.

Then a second error occured because of a bad  400 Bad Request error which was the result of incorrect formatting of the URL after changing FirebaseHistoryRepository and Config to realtime databases.

Then a final error occured when launching the program. The problem was that the MainAppUI was trying to create the DashboardPanel (which contains the HistoryPanel) during the initial UI setup, before any user had logged in. This meant the username was still null when passed to the constructors.